[jscoq] Switch to Web Workers.
JsCoq now runs in a worker and communicates with coq-manager using
messages and PPX serializers.

We have reshaped the API quite a bit. We are very close towards
convergence with SerAPI, from which we have imported quite a few ideas.

The main changes are:

- coq-manager now follows a fully reactive approach (as suggested by
  @cpitclaudel) The document model has also evolved.

  This implies a big improvement in ease of document handling. In
  particular, we do not wait for `Add`, we listen to feedback to color
  states, and we use the new `Cancel` API instead of `EditAt`.

- The library manager is basically shared with serapi, and it has been
  greatly simplified.

- The protocol features a more robust initialization procedure, even
  if a bit heavier on client side responsabilities.

- We now follow Coq 8.7 master (in particular trunk + coq/coq#441).

Many, many hacks have been removed in favor of more principled
solutions; still, we have quite a long way to go.

This commit improves #13, helps a bit with #9 (now we can kill the web
worker), and (in an non-optimal way yet) fixes #16.

Pending items:

+ Pp.t rendering is just a prototype, we need to improve it and insert
  divs following the current box mode.

+ Pp.t normalization leaves a lot to be desired. We can likely
  normalize it Ocaml-side and produce a div. We could even use jsoo's
  Tyxml for that.

- The invalidation, focus, and movement logic needs a bit more of
  work. In particular we need to streamline adding many sentences.